\name{trtf-package}
\docType{package}
\alias{trtf-package}
\title{General Information on the \pkg{trtf} Package}
\description{
  The \pkg{trtf} package implements transformation trees and transformation
  forests as described in Hothorn and Zeileis (2017).

  Example applications of transformation trees and forests can be replicated
  using \code{demo("applications")} and \code{demo("BMI")}. Figure 1
  in Hothorn and Zeileis (2017) can be reproduced by \code{demo("QRF")}.
  Source code of simulation experiments is available in directory 
  \code{trtf/inst/sim}.

  A short talk introducing transformation trees and forests is available from
  \url{https://channel9.msdn.com/Events/useR-international-R-User-conferences/useR-International-R-User-2017-Conference/Transformation-Forests}.

}
\references{

  Torsten Hothorn and Achim Zeileis (2017). Transformation Forests.
  \url{https://arxiv.org/abs/1701.02110}.

}
\author{
  This package is authored by Torsten Hothorn <Torsten.Hothorn@R-project.org>.
}
\keyword{package}
